you now can give an dictonary or a single recording as argument and you get the plots of EODs, PSD and soundtrace saved in figures/

+++ sorting_tools/fish_transect_main.py
@@ -7,23 +7,35 @@ import numpy as np
 from IPython import embed
 
 def main():
-    filepath = sys.argv[1]
-
-    # load txt file; filenames in line
-    if not os.path.exists('files.txt'):
-        os.system('ls %s > files.txt' %filepath)
-
-    txt_file =glob.glob('files.txt')
-    f = open('%s' %txt_file[0], 'r')
+    input = sys.argv[1]
+    file_type = input.split('/')[-1].split('.')[-1]
 
     # temporary file untill we got a file with pulse fundamentals
     if not os.path.exists('fish_pulse.npy'):
         np.save('fish_pulse.npy', np.array([]))
 
-    # for loop with doing till_juan_fishfinder.py with each file
-    for file in f:
-        os.system('python till_juan_fishfinder2.2.py %s%s' %(filepath, file))
-    os.remove('files.txt')
+    if file_type == '':
+        print ''
+        print 'Input is a directory !!!'
+        print ''
+
+        if not os.path.exists('files.txt'):
+            os.system('ls %s > files.txt' %input)
+
+        txt_file =glob.glob('files.txt')
+        f = open('%s' %txt_file[0], 'r')
+
+        # for loop with doing till_juan_fishfinder.py with each file
+        for file in f:
+            os.system('python till_juan_fishfinder2.2.py %s%s' %(input, file))
+        os.remove('files.txt')
+
+    else:
+        print ''
+        print 'Input is a ', file_type, 'file !!!'
+        print ''
+
+        os.system('python till_juan_fishfinder2.2.py %s' %input)
 
     # create_plots.py
     os.system('python create_plots2.py')
